api/requirements.txt
requests
debugpy>=1.0,<2
boto3
jmespath
aws_xray_sdk
aws_lambda_powertools
python-slugify